Replaced JSP server code with Servlet + JS client code.
JS client code is much more portable and efficient than JSP. The use of Servlets (or any internal API) is necessary to protect the corresponding backends (deployer, monitor, sla...) from the Internet. Backend machines should be not exposed to the internet, only the dashboard will be available directly on a public website, dealing with the security.