FEATURE (Dashboard): New version of SeaClouds Dashboard (v2)
This new version is AngularJS based, rewriten from the scratch in order to improve the overall quality of the Dashboard. It follows the design guidelines exposed in D4.5.

Thanks to @rosogon for the work done on the topology editor/viewer and to @mbarrientos for helping me with bugfixes and cosmetic improvements