FEATURE: Adding CF support
CloudFoundry location was added.
A set of entities represent the Applications which can be deployed on CF and the services that can be instantiated and bound to the application.

The CloudFoundry (CF) entities and location are integrated on SeaCloudsPlatform.
An uber jar is generated for Deployer module which contains the required dependencies to manage the CF location and entities.

The deployer jar is integrated as a resource in Installer module, this allows to add the CF features easily the Deployer BrooklynNode.