Added SeaCloudsInitializerPolicy.
This policy is on charge to configure and clean up all the required SeaClouds Components (SLA Core, Tower4Clouds, InfluxDB observers and Grafana Dashboards)  when an application is installed/removed.

This policy is designed to be attached at root level. In order to see working examples please take a look to SeaCloudsInitializerPolicyTest.