Adds specification of frontend node in the designer.
A frontend node is specified in the topology as a boolean property in
the properties of a node. Visually, it appears with black rounded line.
The rest of nodes appears with a grey rounded line. Only one frontend
node can be specified.

This way of specifying a frontend has less priority than the previous
one when generating the AAM (set the frontend node in the
app_requirements element), but it is easier to implement from the designer.

Form code for the topology editor has been moved from editor.js to
editor-forms.js, as the former file was too complex.