Fixes issue #180: Credentials screen not working after new DAM format
In new DAM Generator, PaaS providers do not appear as node templates and
the information should be taken from brooklyn policies.

When generating the topology associated to the DAM, the DAM is now
preprocessed to add PaaSes as node templates in the DAM. This only takes
effect for the generation of the topology, and allows the rest of the
code work as it was.